DataAcquireCode/DataReadout/Python/requirements.txt
bzr==2.8.0.dev1
catkin-pkg==0.4.12
catkin-pkg-modules==0.4.12
chardet==2.3.0
configobj==5.0.6
cryptography==1.7.1
defusedxml==0.4.1
docutils==0.13.1
dronekit==2.9.1
empy==3.3.2
enum34==1.1.6
future==0.17.1
httplib2==0.9.2
idna==2.2
ipaddress==1.0.17
keyring==10.1
keyrings.alt==1.3
launchpadlib==1.10.4
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
lxml==4.3.1
MAVProxy==1.8.5
mercurial==4.0
monotonic==1.2
navio2==1.0.0
netifaces==0.10.4
nose==1.3.7
numpy==1.12.1
oauth==1.0.1
paramiko==2.0.0
Pillow==4.0.0
pyasn1==0.1.9
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.22.0
pygpgme==0.3
pymavlink==2.3.4
pyparsing==2.1.10
pyserial==3.4
Pyste==0.9.10
python-dateutil==2.5.3
pyxdg==0.25
PyYAML==3.12
roman==2.0.0
rosdep==0.15.1
rosdistro==0.7.3
rosdistro-modules==0.7.3
rosinstall==0.7.8
rospkg==1.1.9
rospkg-modules==1.1.9
RPi.GPIO==0.6.5
SecretStorage==2.3.1
simplejson==3.10.0
six==1.12.0
spidev==3.3
vcstools==0.1.40
wadllib==1.3.2
wstool==0.1.17
wxPython==3.0.2.0
wxPython-common==3.0.2.0
zope.interface==4.3.2
